refact: add AbstractCrudServlet
